requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
confluent-kafka==2.4.0
cymem==2.0.8
Flask==3.0.3
flask-restx==1.3.0
flask-swagger-ui==4.11.1
future==1.0.0
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdc==1.2.1
mdurl==0.1.2
murmurhash==1.0.10
numpy==1.23.5
packaging==24.1
pip==24.0
preshed==3.0.9
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
python-json-logger==2.0.7
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.1
uWSGI==2.0.26
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
config==0.5.1
kafka-python==2.0.2
